new dep

+++ contracts/DefifaGovernor.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.16;
 
 import "prb-math/PRBMath.sol";
 
-import "@jbx-protocol/juice-nft-rewards/contracts/JBTiered721Delegate.sol";
-import "@jbx-protocol/juice-nft-rewards/contracts/JB721TieredGovernance.sol";
+import "@jbx-protocol/juice-721-delegate/contracts/JBTiered721Delegate.sol";
+import "@jbx-protocol/juice-721-delegate/contracts/JB721TieredGovernance.sol";
 
 import "@openzeppelin/contracts/governance/Governor.sol";
 import "@openzeppelin/contracts/governance/extensions/GovernorSettings.sol";
@@ -14,12 +14,7 @@ import "@openzeppelin/contracts/governance/extensions/GovernorVotesQuorumFractio
 // errors
 error PROPOSAL_CREATION_THRESHOLD_NOT_REACHED_YET();
 
-contract DefifaGovernor is
-    Governor,
-    GovernorSettings,
-    GovernorCountingSimple
-{
-
+contract DefifaGovernor is Governor, GovernorSettings, GovernorCountingSimple {
     error INCORRECT_TIER_ORDER(uint256, uint256[]);
 
     // The max voting power 1 tier has if everyone votes
@@ -33,7 +28,7 @@ contract DefifaGovernor is
 
     uint256 internal immutable DEPLOYED_AT;
 
-   // datasource
+    // datasource
     IJBTiered721Delegate jbTieredRewards;
 
     // The datasource for votingpower
@@ -69,26 +64,26 @@ contract DefifaGovernor is
         bytes memory params
     ) internal view virtual override(Governor) returns (uint256 votingPower) {
         // Decode the bytes into the tier_ids
-        uint256[] memory _tiers = abi.decode(params, (uint[]));
+        uint256[] memory _tiers = abi.decode(params, (uint256[]));
         uint256 _tiers_length = _tiers.length;
 
         // Loop over all tiers gathering the voting share of the user
         uint256 _prevTier;
         for (uint256 _i; _i < _tiers_length; ) {
-            // Enforce the tiers to be in ascending order, reverts if 
+            // Enforce the tiers to be in ascending order, reverts if
             // there are any duplicates or the tiers are incorrecly sorted
-            if(_tiers[_i] <= _prevTier) revert INCORRECT_TIER_ORDER(_i, _tiers);
+            if (_tiers[_i] <= _prevTier)
+                revert INCORRECT_TIER_ORDER(_i, _tiers);
             _prevTier = _tiers[_i];
 
             uint256 _tierVotingPower = jbTieredGovernance.getPastTierVotes(
-
                 account,
                 _tiers[_i],
                 blockNumber
             );
 
             unchecked {
-                if (_tierVotingPower != 0){
+                if (_tierVotingPower != 0) {
                     votingPower += PRBMath.mulDiv(
                         _tierVotingPower,
                         MAX_VOTING_POWER_TIER,
@@ -126,19 +121,21 @@ contract DefifaGovernor is
         override
         returns (bytes memory)
     {
-       // TODO: should we do this on every time or should we just store this, what is cheaper...
-       uint256 _count = jbTieredRewards.store().maxTierIdOf(address(jbTieredRewards));
-       uint256[] memory _ids = new uint256[](_count);
+        // TODO: should we do this on every time or should we just store this, what is cheaper...
+        uint256 _count = jbTieredRewards.store().maxTierIdOf(
+            address(jbTieredRewards)
+        );
+        uint256[] memory _ids = new uint256[](_count);
 
-      // Add all tiers to the array
-      for(uint256 _i; _i < _count;){
-          // Tiers start counting from 1
-         _ids[_i] = _i + 1;
+        // Add all tiers to the array
+        for (uint256 _i; _i < _count; ) {
+            // Tiers start counting from 1
+            _ids[_i] = _i + 1;
 
-         unchecked{ 
-            ++_i;
-         }
-      }
+            unchecked {
+                ++_i;
+            }
+        }
 
         return abi.encode(_ids);
     }
@@ -152,8 +149,13 @@ contract DefifaGovernor is
         returns (uint256)
     {
         // After the contract initially deploys there is a long delay, once this long delay has passed we use `VOTING_DELAY`
-        if(DEPLOYED_AT + INITIAL_VOTING_DELAY_AFTER_DEPLOYMENT - VOTING_DELAY > block.timestamp){
-            return ((DEPLOYED_AT + INITIAL_VOTING_DELAY_AFTER_DEPLOYMENT) - block.timestamp) / BLOCKTIME_SECONDS;
+        if (
+            DEPLOYED_AT + INITIAL_VOTING_DELAY_AFTER_DEPLOYMENT - VOTING_DELAY >
+            block.timestamp
+        ) {
+            return
+                ((DEPLOYED_AT + INITIAL_VOTING_DELAY_AFTER_DEPLOYMENT) -
+                    block.timestamp) / BLOCKTIME_SECONDS;
         }
 
         return VOTING_DELAY / BLOCKTIME_SECONDS;
@@ -226,7 +228,6 @@ contract DefifaGovernor is
         revert("use castVote");
     }
 
-
     /**
      * @dev See {IGovernor-castVoteWithReasonAndParamsBySig}.
      * reverting to avoid passsing of duplicate tier id's in params
@@ -283,12 +284,7 @@ contract DefifaGovernor is
         return super._cancel(targets, values, calldatas, descriptionHash);
     }
 
-    function _executor()
-        internal
-        view
-        override(Governor)
-        returns (address)
-    {
+    function _executor() internal view override(Governor) returns (address) {
         return super._executor();
     }
